% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvar.R
\name{envvar_proxy}
\alias{envvar_proxy}
\alias{envvar_no_proxy}
\title{Get environment-variables for proxy}
\usage{
envvar_proxy()

envvar_no_proxy()
}
\value{
\code{character} value of proxy or no_proxy environment-variable.
}
\description{
\describe{
\item{\code{envvar_proxy()}}{returns first value found in the sequence of
environment variables:
\code{"http_proxy"}, \code{"HTTP_PROXY"}, \code{"https_proxy"}, \code{"HTTPS_PROXY"}.
}
\item{\code{envvar_no_proxy()}}{returns first value found in the sequence of
environment variables: \code{"no_proxy"}, \code{"NO_PROXY"}.}
}
}
\details{
Note that \code{envvar_proxy()} returns \code{""} for an unset environment-variable.
This is different from \code{\link[=get_proxy]{get_proxy()}}, which will return \code{NA} if the proxy
should not be used. This reflects the difference in syntax between
between \code{\link[=Sys.getenv]{Sys.getenv()}} and \code{\link[withr:with_envvar]{withr::with_envvar()}} .
}
\examples{
  envvar_proxy()
  envvar_no_proxy()
}
