% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-proxy.R
\name{local_proxy_for_url}
\alias{local_proxy_for_url}
\alias{with_proxy_for_url}
\alias{without_proxy}
\title{Set proxy environment-variables temporarily}
\usage{
local_proxy_for_url(url, .local_envir = parent.frame())

with_proxy_for_url(url, code)

without_proxy(code)
}
\arguments{
\item{url}{\code{character} address for which to determine the proxy.}

\item{.local_envir}{\code{environment}, the environment to use for scoping.}

\item{code}{The code to execute in the temporary environment.}
}
\value{
These functions are called for their side-effects:
\describe{
\item{\code{local_proxy_for_url()}}{\code{character} vector of previous values of
proxy environment-variables, invisibly.}
\item{\code{with_proxy_for_url()}}{The results of the evaluation of the \code{code}
argument.}
\item{\code{without_proxy()}}{The results of the evaluation of the \code{code}
argument.}
}
}
\description{
Given a \code{url} and the your current proxy and no-proxy environment-variables,
these functions \emph{temporarily} set your proxy environment-variables:
\describe{
\item{\code{local_proxy_for_url()}}{Set until the current evaluation context
ends (such as the end of a function).}
\item{\code{with_proxy_for_url()}}{Set only for the execution of the
enclosed \code{code}.}
\item{\code{without_proxy()}}{For the execution of the enclosed \code{code},
regardless of \code{url}, unsets the proxy environment-variables.}
}

These functions \code{local_proxy_for_url()} and \code{with_proxy_for_url()} are useful
in situations where the proxy variables are effective, and the no-proxy
variable is not effective.

The function \code{without_proxy()} particularly useful for when you need to call
\emph{another} function interactively, but need to turn the proxy off.
}
\details{
These functions themselves call \link[withr:withr]{withr::withr} \code{with_} and \code{local_} functions;
they behave in the same way as the withr functions.
}
\examples{
# sets your proxy variables to the right value for a given `url`
local_proxy_for_url("https://github.acme.com")

# same as above, but for enclosed code
with_proxy_for_url(
  "https://github.acme.com",
  print("you will likely run more interesting code than this")
)

# run with proxy off
without_proxy(
  print("anything else is more interesting than this.")
)
}
