% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-proxy.R
\name{uses_proxy}
\alias{uses_proxy}
\alias{get_proxy}
\title{Determine proxy}
\usage{
uses_proxy(url, no_proxy = envvar_no_proxy())

get_proxy(url, proxy = envvar_proxy(), no_proxy = envvar_no_proxy())
}
\arguments{
\item{url}{\code{character} address for which to determine the proxy.}

\item{no_proxy}{\code{character} comma-separated string containing hostnames
where the proxy is not to be used.}

\item{proxy}{\code{character} the proxy that would be used.}
}
\value{
\describe{
\item{\code{uses_proxy()}}{\code{logical} indicating if the proxy should be used.}
\item{\code{get_proxy()}}{\code{character} value of proxy to use.}
}
}
\description{
Use these functions to determine if a \code{proxy} should be used for a given
\code{url} and \code{no_proxy} string.
}
\details{
This aims to emulate the
\href{https://curl.se/libcurl/c/CURLOPT_NOPROXY.html}{cURL specification}
for \code{NO_PROXY}. The biggest difference between this implementation and the
cURL specification is that the wildcard character (\code{*}) is used more
permissably here.
}
\examples{
  proxy <- "http://proxy.acme.com"
  no_proxy <- "github.acme.com,localhost,127.0.0.1"

  uses_proxy("https://github.com", no_proxy)      # TRUE
  uses_proxy("https://github.acme.com", no_proxy) # FALSE

  get_proxy("https://github.com", proxy, no_proxy) # "http://proxy.acme.com"
  get_proxy("https://github.acme.com", proxy, no_proxy) # ""
}
