% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-proxy.R
\name{uses_proxy}
\alias{uses_proxy}
\alias{get_proxy}
\title{Determine proxy}
\usage{
uses_proxy(url, no_proxy = envvar_no_proxy())

get_proxy(url, proxy = envvar_proxy(), no_proxy = envvar_no_proxy())
}
\arguments{
\item{url}{\code{character}}

\item{no_proxy}{\code{character}}

\item{proxy}{\code{character}}
}
\value{
\describe{
\item{\code{uses_proxy()}}{\code{logical} indicating if the proxy should be used.}
\item{\code{get_proxy()}}{\code{character} value of proxy to use.}
}
}
\description{
Determine proxy
}
\examples{
  proxy <- "http://proxy.acme.com"
  no_proxy <- "github.acme.com,localhost,127.0.0.1"

  uses_proxy("https://github.com", no_proxy)      # TRUE
  uses_proxy("https://github.acme.com", no_proxy) # FALSE

  get_proxy("https://github.com", proxy, no_proxy) # "http://proxy.acme.com"
  get_proxy("https://github.acme.com", proxy, no_proxy) # ""
}
